refactor: 템플릿 생성 로직 추상화
