+++ index.html
@@ -28,7 +28,7 @@
             </nav>
           </header>
           <main class="mt-10">
-            <section id="video-list" class="video-wrapper"></section>
+            <section id="video-list"></section>
           </main>
         </div>
       </div>
